refactor: move Input/Result View
